[style] 상세 페이지 버튼 컴포넌트
