+++ my-app/src/styles/GlobalStyle.tsx
@@ -13,6 +13,7 @@ const GlobalStyle = createGlobalStyle`
 
   a {
       text-decoration: none;
+      color: black;
     }
 
   ul, li {
